Added audience validation check to protect against a confused deputy attack.
This is an attack where the publisher of another app has a legitimately obtained access token for the user for *their* app, and attempts to use that access token to access our app on behalf of that user. Checking the audience ensures that only tokens issued to the user in the context of *our* app can be used.